requirements.txt
yt-dlp>=2024.1.0
openai>=1.0.0
python-dotenv>=1.0.0
rich>=13.0.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
